Use long instead of int.
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/collections/trunk@1682774 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/commons/collections4/IterableUtils.java
@@ -213,7 +213,7 @@ public class IterableUtils {
      * @return a bounded view on the specified iterable
      * @throws IllegalArgumentException if maxSize is negative
      */
-    public static <E> Iterable<E> boundedIterable(final Iterable<E> iterable, final int maxSize) {
+    public static <E> Iterable<E> boundedIterable(final Iterable<E> iterable, final long maxSize) {
         if (maxSize < 0) {
             throw new IllegalArgumentException("maxSize parameter must not be negative.");
         }
@@ -311,7 +311,7 @@ public class IterableUtils {
      * @return a view of the specified iterable, skipping the first N elements
      * @throws IllegalArgumentException if elementsToSkip is negative
      */
-    public static <E> Iterable<E> skippingIterable(final Iterable<E> iterable, final int elementsToSkip) {
+    public static <E> Iterable<E> skippingIterable(final Iterable<E> iterable, final long elementsToSkip) {
         if (elementsToSkip < 0) {
             throw new IllegalArgumentException("elementsToSkip parameter must not be negative.");
         }
